Implemented PR comments
