+++ prediction_market_agent/agents/microchain_agent/functions.py
@@ -19,6 +19,7 @@ from prediction_market_agent_tooling.tools.balances import get_balances
 
 from prediction_market_agent.agents.microchain_agent.utils import (
     MicroMarket,
+    fetch_public_key_from_env,
     get_omen_binary_market_from_question,
     get_omen_binary_markets,
     get_omen_market_token_balance,
@@ -106,6 +107,7 @@ class GetBalance(Function):
 
 class BuyTokens(Function):
     def __init__(self, outcome: str):
+        self.user_address = fetch_public_key_from_env()
         self.outcome = outcome
         super().__init__()
 
@@ -121,14 +123,23 @@ class BuyTokens(Function):
         outcome_bool = get_boolean_outcome(self.outcome)
 
         market_obj: OmenAgentMarket = get_omen_binary_market_from_question(market)
+        outcome_index = market_obj.get_outcome_index(self.outcome)
+        market_index_set = outcome_index + 1
+
         before_balance = get_omen_market_token_balance(
-            market=market_obj, outcome=outcome_bool
+            user_address=self.user_address,
+            market_condition_id=market_obj.condition.id,
+            market_index_set=market_index_set,
         )
         market_obj.place_bet(
             outcome_bool, BetAmount(amount=Decimal(amount), currency=Currency.xDai)
         )
         tokens = (
-            get_omen_market_token_balance(market=market_obj, outcome=outcome_bool)
+            get_omen_market_token_balance(
+                user_address=self.user_address,
+                market_condition_id=market_obj.condition.id,
+                market_index_set=market_index_set,
+            )
             - before_balance
         )
         return f"Bought {tokens} {self.outcome} outcome tokens of: {market}"
@@ -243,12 +254,9 @@ class GetUserPositions(Function):
         return ["0x2DD9f5678484C1F59F97eD334725858b938B4102"]
 
     def __call__(self, user_address: str) -> list[OmenUserPosition]:
-        user_address_checksummed = to_checksum_address(user_address)
-        omen_subgraph_handler = OmenSubgraphHandler()
-        user_positions = omen_subgraph_handler.get_user_positions(
-            better_address=user_address_checksummed
+        return OmenSubgraphHandler().get_user_positions(
+            better_address=to_checksum_address(user_address)
         )
-        return user_positions
 
 
 ALL_FUNCTIONS = [
